requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
ffmpeg-python==0.2.0
filelock==3.12.3; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
future==0.18.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4; python_version >= '3.5'
jinja2==3.1.2; python_version >= '3.7'
llvmlite==0.40.1; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
more-itertools==10.1.0; python_version >= '3.8'
mpmath==1.3.0
multidict==6.0.4; python_version >= '3.7'
networkx==3.1; python_version >= '3.8'
numba==0.57.1; python_version >= '3.8'
numpy==1.24.4; python_version >= '3.8'
openai==0.28.0; python_full_version >= '3.7.1'
openai-whisper==20230314; python_version >= '3.8'
pydub==0.25.1
regex==2023.8.8; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
sympy==1.12; python_version >= '3.8'
tiktoken==0.3.1; python_version >= '3.8'
torch==2.0.1; python_full_version >= '3.8.0'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.7.1; python_version >= '3.7'
urllib3==2.0.4; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
